requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
httpx
python-multipart
